Added Discard Button
Change Log:
- Changed card names and descriptions to overall shorten them, and to have consistent terminology.
-  Changed some card values.
- Added a Bin Button, which allows the player to discard the selected card, and gain energy equal to half of the card's cost.
- The Bless effect is now called Boost.

+++ Assets/Scripts/Managers/HandManager.cs
@@ -23,7 +23,7 @@ public class HandManager : MonoBehaviour
     [SerializeField] public List<int> deckCardIDs = new List<int>();
     [SerializeField] public List<int> handCardIDs = new List<int>();
     public int drawNum = 3; // Equal to initial hand size
-    public int actionPoints = 2; // 3 for each unit:0-Bury Card, 1-Draw Random, 2-Shuffle & Redraw Hand, 3-Draw Specific Card
+    public float actionPoints = 2; // 3 for each unit:0-Bury Card, 1-Draw Random, 2-Shuffle & Redraw Hand, 3-Draw Specific Card
     [SerializeField] public int deckIndex = 0; // Points to card to draw from deck
     public bool handDrawn = false;
     //public bool handSelected = false;
